GamePage broken down into components
